use bigint for detections; use timescale for test; fix tests re #431

+++ common/testing/sequelize.js
@@ -1,13 +1,10 @@
 const fs = require('fs')
 const path = require('path')
 const express = require('express')
-const excludedMigrations = [
-  '20220421000001-enable-auto-increment-users.js'
-]
 
 // Copied from StackOverflow - only to be used for testing
-async function migrate (sequelize, Sequelize, table = '`SequelizeMeta`') {
-  const migrations = fs.readdirSync(path.join(__dirname, '../../core/_cli/migrations')).filter(filePath => filePath.endsWith('.js')).filter(filePath => !excludedMigrations.includes(filePath))
+async function migrate (sequelize, Sequelize, table = '"SequelizeMeta"') {
+  const migrations = fs.readdirSync(path.join(__dirname, '../../core/_cli/migrations')).filter(filePath => filePath.endsWith('.js'))
   await sequelize.query(`CREATE TABLE IF NOT EXISTS ${table} (name VARCHAR(255) NOT NULL UNIQUE)`)
   const completedMigrations = await sequelize.query(`SELECT * FROM ${table}`, { type: Sequelize.QueryTypes.SELECT })
 
@@ -22,12 +19,7 @@ async function migrate (sequelize, Sequelize, table = '`SequelizeMeta`') {
   }
 
   const query = sequelize.queryInterface.sequelize.query
-  const regex = /(create_hypertable|CREATE INDEX|ADD CONSTRAINT|DROP CONSTRAINT|DELETE FROM .* USING)/ // unsupported by sqlite
   sequelize.queryInterface.sequelize.query = (sql, options = {}) => {
-    if (regex.test(sql)) {
-      // console.error('Skip unsupported query: ' + sql)
-      return Promise.resolve()
-    }
     return query.call(sequelize.queryInterface.sequelize, sql, options)
   }
 
@@ -37,7 +29,7 @@ async function migrate (sequelize, Sequelize, table = '`SequelizeMeta`') {
       await migration.up(sequelize.queryInterface, Sequelize)
       await sequelize.query(`INSERT INTO ${table} VALUES (:name)`, { type: Sequelize.QueryTypes.INSERT, replacements: { name: filename } })
     } catch (err) {
-      console.error('Failed performing migration: ' + filename)
+      console.error(`Failed performing migration: ${filename}`, err)
       break
     }
   }
@@ -46,17 +38,17 @@ async function migrate (sequelize, Sequelize, table = '`SequelizeMeta`') {
 }
 
 const primaryUserId = 1
-const primaryUserGuid = 'abc123'
+const primaryUserGuid = 'b7cc2c4e-bea3-11ed-afa1-0242ac120001'
 const primaryUserEmail = 'jb@astonmartin.com'
 const primaryUserFirstname = 'James'
 const primaryUserLastname = 'Bond'
 const otherUserId = 2
-const otherUserGuid = 'def456'
+const otherUserGuid = 'b7cc2c4e-bea3-11ed-afa1-0242ac120002'
 const otherUserEmail = 'em@astonmartin.com'
 const otherUserFirstname = 'Eve'
 const otherUserLastname = 'Moneypenny'
 const anotherUserId = 3
-const anotherUserGuid = 'ghy789'
+const anotherUserGuid = 'b7cc2c4e-bea3-11ed-afa1-0242ac120003'
 const anotherUserEmail = 'big@bobby.com'
 const anotherUserFirstname = 'Big'
 const anotherUserLastname = 'Bobby'
@@ -71,40 +63,51 @@ const seedValues = {
   primaryUserLastname,
   otherUserId,
   otherUserGuid,
+  otherUserEmail,
   anotherUserId,
   anotherUserGuid,
+  anotherUserEmail,
   roleAdmin,
   roleMember,
   roleGuest
 }
 
 async function seed (models) {
-  await models.User.create({ id: primaryUserId, guid: primaryUserGuid, username: 'jb', firstname: primaryUserFirstname, lastname: primaryUserLastname, email: primaryUserEmail })
-  await models.User.create({ id: otherUserId, guid: otherUserGuid, username: 'em', firstname: otherUserFirstname, lastname: otherUserLastname, email: otherUserEmail })
-  await models.User.create({ id: anotherUserId, guid: anotherUserGuid, username: 'st', firstname: anotherUserFirstname, lastname: anotherUserLastname, email: anotherUserEmail })
-  await models.Role.create({ id: roleAdmin, name: 'Admin' })
-  await models.Role.create({ id: roleMember, name: 'Member' })
-  await models.Role.create({ id: roleGuest, name: 'Guest' })
-  await models.RolePermission.create({ role_id: roleAdmin, permission: 'C' })
-  await models.RolePermission.create({ role_id: roleAdmin, permission: 'R' })
-  await models.RolePermission.create({ role_id: roleAdmin, permission: 'U' })
-  await models.RolePermission.create({ role_id: roleAdmin, permission: 'D' })
-  await models.RolePermission.create({ role_id: roleMember, permission: 'C' })
-  await models.RolePermission.create({ role_id: roleMember, permission: 'R' })
-  await models.RolePermission.create({ role_id: roleMember, permission: 'U' })
-  await models.RolePermission.create({ role_id: roleGuest, permission: 'R' })
-  await models.ClassificationSource.create({ id: 1, value: 'unknown' })
-  await models.ClassificationType.create({ id: 1, value: 'unknown' })
+  await models.User.findOrCreate({ where: { id: primaryUserId, guid: primaryUserGuid, username: 'jb', firstname: primaryUserFirstname, lastname: primaryUserLastname, email: primaryUserEmail } })
+  await models.User.findOrCreate({ where: { id: otherUserId, guid: otherUserGuid, username: 'em', firstname: otherUserFirstname, lastname: otherUserLastname, email: otherUserEmail } })
+  await models.User.findOrCreate({ where: { id: anotherUserId, guid: anotherUserGuid, username: 'st', firstname: anotherUserFirstname, lastname: anotherUserLastname, email: anotherUserEmail } })
+  await models.Role.findOrCreate({ where: { id: roleAdmin, name: 'Admin' } })
+  await models.Role.findOrCreate({ where: { id: roleMember, name: 'Member' } })
+  await models.Role.findOrCreate({ where: { id: roleGuest, name: 'Guest' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleAdmin, permission: 'C' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleAdmin, permission: 'R' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleAdmin, permission: 'U' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleAdmin, permission: 'D' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleMember, permission: 'C' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleMember, permission: 'R' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleMember, permission: 'U' } })
+  await models.RolePermission.findOrCreate({ where: { role_id: roleGuest, permission: 'R' } })
+  await models.ClassificationSource.findOrCreate({ where: { id: 1, value: 'unknown' } })
+  await models.ClassificationType.findOrCreate({ where: { id: 1, value: 'unknown' } })
 }
 
 const truncateOrder = ['Event', 'Annotation', 'Detection', 'ClassifierJob', 'ClassifierActiveProject', 'ClassifierActiveStream', 'ClassifierDeployment', 'ClassifierEventStrategy', 'ClassifierOutput', 'Classifier', 'EventStrategy', 'ClassificationAlternativeName', 'Classification', 'UserStreamRole', 'UserProjectRole', 'UserOrganizationRole', 'StreamSegment', 'StreamSourceFile', 'AudioFileFormat', 'AudioCodec', 'FileExtension', 'Stream', 'Project', 'Organization']
 
 async function truncate (models) {
-  return await Promise.all(
-    truncateOrder.map(key => {
-      return models[key].destroy({ where: {}, force: true })
-    })
-  )
+  for (const key of truncateOrder) {
+    if (models[key] !== undefined) {
+      await models[key].destroy({ where: {}, force: true })
+    }
+  }
+}
+
+async function truncateNonBase (models) {
+  const m = Object.assign({}, models)
+  const baseSeed = ['User', 'RolePermission', 'ClassificationSource', 'ClassificationType']
+  baseSeed.forEach((k) => {
+    delete m[k]
+  })
+  return await truncate(m)
 }
 
 function expressApp (userAdditions = {}) {
@@ -126,4 +129,4 @@ function muteConsole (levels = ['log', 'info', 'warn', 'error']) {
   })
 }
 
-module.exports = { migrate, seed, seedValues, truncate, expressApp, muteConsole }
+module.exports = { migrate, seed, seedValues, truncate, truncateNonBase, expressApp, muteConsole }
